requirements.txt
# Core
torch>=1.13.0
torchvision>=0.14.0
torchaudio>=0.13.0

# Scientific
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.2.0
pandas>=2.0.0

# Vision / face
opencv-python>=4.7.0
Pillow>=9.5.0
albumentations>=1.3.0

# Metrics
pingouin>=0.5.3          # ICC
scikit-learn>=1.2.0      # ECE via calibration_curve

# Config
PyYAML>=6.0
omegaconf>=2.3.0

# Logging
tensorboard>=2.13.0
tqdm>=4.65.0
rich>=13.3.0

# Federated / privacy
opacus>=1.4.0            # DP-SGD

# Physiology features
neurokit2>=0.2.7         # HRV feature extraction
heartpy>=1.2.7           # alternative HRV

# Edge / optimisation (optional)
# tensorrt                # uncomment for Jetson Nano TensorRT export
onnx>=1.14.0
onnxruntime>=1.14.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0
